+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- `extensions` and `vtex*` accounts can submit `vtex` apps as allowed first party apps vendors
+
 ## [0.1.2] - 2020-10-07
 ### Fixed
 - Fix relative imports